ShallowWaterModels.jl

ShallowWaterModels.BellCurve — Type

BellCurve(param)
param should contain a value theta

\[h = 2^{(-|x|^\theta)}\]

\[u = 0\]

ShallowWaterModels.Boussinesq — Type

Boussinesq(params)
A Hamiltonian Boussinesq system derived in BonaSmith76 and studied in BonaChenSaut02

ShallowWaterModels.CGBSW — Type

CGBSW( params )

ShallowWaterModels.CGBSW_naive — Type

CGBSW_naive( params )

ShallowWaterModels.HighFreq — Type

HighFreq(param)
param should contain s (regularity index) and freq (frequencies)

ShallowWaterModels.Matsuno_mod_naive — Type

Modified Matsuno models with a naive step function
Matsuno_mod_naive(params)

ShallowWaterModels.Matsuno_naive — Type

Matsuno(params)

ShallowWaterModels.Random — Type

Random(param,s,k)
param should contain s (regularity index)

ShallowWaterModels.fdBoussinesq_1 — Type

fdBoussinesq_1(params)
This model has been introduced and studied by E. Dinvay and collaborators

ShallowWaterModels.fdBoussinesq_1b — Type

fdBoussinesq_1(params)
This model has been introduced and studied by E. Dinvay and collaborators

ShallowWaterModels.fdBoussinesq_2 — Type

fdBoussinesq_1(params)
A quasilinear version of the model introduced and studied by E. Dinvay and collaborators

ShallowWaterModels.fdBoussinesq_2b — Type

fdBoussinesq_2(params)
A quasilinear version of the model introduced and studied by E. Dinvay and collaborators

ShallowWaterModels.mapfro — Method

mapfro(Boussinesq, data)

ShallowWaterModels.mapfro — Method

mapfro(CGBSW, data)

ShallowWaterModels.mapfro — Method

mapfro(CGBSW_naive, data)

ShallowWaterModels.mapfro — Method

mapfro(Matsuno, data)

ShallowWaterModels.mapfro — Method

mapfro(Matsuno, data)

ShallowWaterModels.mapfro — Method

mapfro(Matsuno, data)

ShallowWaterModels.mapfro — Method

mapfro(fdBoussinesq_1, data)

ShallowWaterModels.mapfro — Method

mapfro(fdBoussinesq_1b, data)

ShallowWaterModels.mapfro — Method

mapfro(fdBoussinesq_2, data)

ShallowWaterModels.mapfro — Method

mapfro(fdBoussinesq_2b, data)

ShallowWaterModels.mapto — Method

mapto(Boussinesq, data)

ShallowWaterModels.mapto — Method

mapto(CGBSW, data)

ShallowWaterModels.mapto — Method

mapto(CGBSW_naive, data)

ShallowWaterModels.mapto — Method

mapto(Matsuno, data)

ShallowWaterModels.mapto — Method

mapto(Matsuno, data)

ShallowWaterModels.mapto — Method

mapto(Matsuno, data)

ShallowWaterModels.mapto — Method

mapto(fdBoussinesq_1, data)

ShallowWaterModels.mapto — Method

mapto(fdBoussinesq_1b, data)

ShallowWaterModels.mapto — Method

mapto(fdBoussinesq_2, data)

ShallowWaterModels.mapto — Method

mapto(fdBoussinesq_2b, data)