Code basics

Abstract types

TimeSolver (RK4, Euler, etc),

AbstractModel (Cheng, Matsuno, etc),

InitialData (Bump, SolitaryWave, etc)

Instances are created from Parameters type.

Missing docstring.

Missing docstring for Parameters. Check Documenter's build log for details.

Une structure Problem représente un problème donné que l'on va résoudre. Les données seront stockées dans data, qui est vide initialement.

ShallowWaterModels.Problem — Type

Problem( model, initial, param, solver)

model : CGBSW or Matsuno

initial : BellCurve

param : must contain N, L, T, dt for Mesh and Times, may contain additional data for Models (ϵ)

solver : RK4 (optional)

Initial data

Missing docstring.

Missing docstring for Bump. Check Documenter's build log for details.

Shallow water models

Missing docstring.

Missing docstring for Cheng. Check Documenter's build log for details.

ShallowWaterModels.Matsuno — Type

Matsuno(params)

Missing docstring.

Missing docstring for solve!(::Problem). Check Documenter's build log for details.

Main program

epsilon = 1/2
N       = 2^12
L       = 10
T       = 5
dt      = 0.001

param = Parameters(epsilon,N,L,T,dt)

problems = [Problem(Cheng,Bump,param,RK4),Problem(Matsuno,Bump,param,RK4)]

Cheng est le modèle utilisé. Il prend en valeur Bump, param et définit 3 fonction:

init qui construit une donnée initiale à partir de Bump : Uinit=init(Bump,param)

Fwave utilisée pour résoudre dt U= Fwave(U) (avec donnée initiale Uinit)

build qui reconstruit la donnée finale (c'est l'application inverse de init). Ufin=final(U,param)

Missing docstring.

Missing docstring for init(::Cheng). Check Documenter's build log for details.

Missing docstring.

Missing docstring for init(::Matsuno). Check Documenter's build log for details.

Missing docstring.

Missing docstring for build(::Cheng). Check Documenter's build log for details.

Missing docstring.

Missing docstring for build(::Matsuno). Check Documenter's build log for details.

simuls = [(Cheng,Bump,param1,RK4),(Cheng,Bump,param2,RK4)]

RK4 solver

ShallowWaterModels.RK4 — Type

RK4(params)

Runge-Kutta fourth order solver.

for p in problems
    solve!(p)
end