Graphics

WaterWaves1D.CnoidalWaveSerreGreenNaghdi — Method

CnoidalWaveSerreGreenNaghdi(param; P=1)

Compute the Serre-Green-Naghdi cnoidal wave with prescribed h₀<h₁<h₂. h₁ is the minimum, h₂ is the maximum of the wave. As h₀ -> h₁, the cnoidal wave converges towards the solitary wave. See for instance Gavrilyuk, Nkonga, Shyue and Truskinovsky, doi:10.1088/1361-6544/ab95ac

Arguments

param :: NamedTuple: parameters of the problem containing h₀<h₁<h₂ and dimensionless parameters ϵ and μ, and number of collocation points N.

P :: Int: (keyword, optional, default = 1) the number of periods of the cnoidal wave in the constructed mesh.

Return values

(η,u,v,mesh,param) with

η :: Vector{Float64}: surface deformation;

u :: Vector{Float64}: layer-averaged velocity;

v :: Vector{Float64}: derivative of the trace of the velocity potential at the surface;

mesh :: Mesh: mesh collocation points;

param :: NamedTuple: useful parameters

WaterWaves1D.SolitaryWaveSerreGreenNaghdi — Method

SolitaryWaveSerreGreenNaghdi(param; x₀=0)

Compute the Serre-Green-Naghdi solitary wave with prescribed velocity.

Arguments

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ, and mesh size L and number of collocation points N;

x₀ :: Real: (keyword, optional, default = 0) center of solitary wave.

Return values

(η,u,v,mesh) with

η :: Vector{Float64}: surface deformation;

u :: Vector{Float64}: layer-averaged velocity;

v :: Vector{Float64}: derivative of the trace of the velocity potential at the surface;

mesh :: Mesh: mesh collocation points.

WaterWaves1D.SolitaryWaveWhitham — Method

SolitaryWaveWhitham(param; kwargs)

Compute the Whitham solitary wave with prescribed velocity.

Argument

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ, and mesh size L and number of collocation points N;

Keywords (optional)

guess :: Vector{Real}: initial guess for the surface deformation (if not provided, the exact formula for KdV is used);

x₀ :: Real: center of solitary wave (if guess is not provided);

iterative :: Bool: inverts Jacobian through GMRES if true, LU decomposition if false;

verbose :: Bool: prints numerical errors at each step if true;

max_iter :: Int: maximum number of iterations of the Newton algorithm;

tol :: Real: general tolerance (default is 1e-10);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

gtol :: Real: relative tolerance of the GMRES algorithm (default is 1e-10);

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

q :: Real: Newton algorithm modified with

u_{n+1}=q*u_{n+1}+(1-q)*u_n (default is 1);

α :: Real: adds α times spectral projection onto the Kernel to the Jacobian;

KdV :: Bool: if true computes the KdV (instead of Whitham) solitary wave.

Return values

(u,mesh) with

u :: Vector{Float64}: the solution;

mesh :: Mesh: mesh collocation points.

WaterWaves1D.SolitaryWaveWhithamBoussinesq — Method

SolitaryWaveWhithamBoussinesq(param; kwargs)

Compute the Whitham-Boussinesq solitary wave with prescribed velocity.

Argument

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ,

and mesh size L and number of collocation points N;

Keywords (optional)

guess :: Vector{Real}: initial guess for the surface deformation (if not provided, the exact formula for SGN is used);

x₀ :: Real: center of solitary wave (if guess is not provided);

α :: Real: determines the model used (typically 1 or 1/2, default is 1);

Boussinesq: if true (default is false), compute the standard Boussinesq system with parameters a (defaut -1//3), b=d (defaut 1//3), and c=0);

iterative :: Bool: inverts Jacobian through GMRES if true, LU decomposition if false;

verbose :: Bool: prints numerical errors at each step if true;

max_iter :: Int: maximum number of iterations of the Newton algorithm;

tol :: Real: general tolerance (default is 1e-10);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

gtol :: Real: relative tolerance of the GMRES algorithm (default is 1e-10);

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

q :: Real: Newton algorithm modified with

u_{n+1}=q*(u_n+du)+(1-q)*u_n (default is 1);

β :: Real: adds β times spectral projection onto the Kernel to the Jacobian.β

Return values

(η,v,mesh) with

η :: Vector{Float64}: surface deformation;

v :: Vector{Float64}: velocity (derivative of the trace of the velocity potential at the surface);

mesh :: Mesh: mesh collocation points.

WaterWaves1D.SolitaryWaveWhithamGreenNaghdi — Method

SolitaryWaveWhithamGreenNaghdi(param; kwargs)

Compute the Whitham-Green-Naghdi solitary wave with prescribed velocity.

Arguments

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ, and mesh size L and number of collocation points N;

Keywords (optional)

guess :: Vector{Real}: initial guess for the surface deformation (if not provided, the exact formula for SGN is used);

x₀ :: Real: center of solitary wave (if guess is not provided);

SGN :: Bool: if true computes the Serre-Green-Naghdi (instead of Whitham-Green-Naghdi) solitary wave (consider SolitaryWaveSerreGreenNaghdi instead);

method :: Int: equation used (between 1 and 4);

iterative :: Bool: inverts Jacobian through GMRES if true, LU decomposition if false (default is false);

verbose :: Bool: prints numerical errors at each step if true (default is false);

max_iter :: Int: maximum number of iterations of the Newton algorithm (default is 20);

tol :: Real: relative tolerance measured in ℓ∞ norm (default is 1e-10);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

gtol :: Real: relative tolerance of the GMRES algorithm (default is 1e-10);

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

q :: Real: Newton algorithm modified with u_{n+1}=q*u_{n+1}+(1-q)*u_n (default is 1);

α :: Real: adds α times spectral projection onto the Kernel to the Jacobian (default is 0).

Return values

(η,u,v,mesh) with

η :: Vector{Float64}: surface deformation;

u :: Vector{Float64}: layer-averaged velocity;

v :: Vector{Float64}: derivative of the velocity potential at the surface;

mesh :: Mesh: mesh collocation points.

WaterWaves1D.create_animation — Method

create_animation( problems; name::String, kwargs... )

Create an animation showing the evolution of initial-value problems.

Argument problems is either an element or a collection (vector, list, etc.) of elements of type Problem.

The animation is saved as name.gif if name is provided.

Other keyword arguments are as follows

xlims allows to specifies the x axis limits for the surface deformation. If nothing is provided (default), then the full numerical basin is represented.

ylims allows to specifies the y axis limits for the surface deformation. If nothing is provided (default), then the limits are determined from the initial data. If anything but a Tuple is provided, the axis limits evolve with the solution.

vlims and flims are as above, but for the velocity and Fourier coefficients plots.

Nframes gives the (maximal) number of frames in the animation.

other arguments of plot_solution!

Return anim, an animation, which can then generate (for instance) a gif through gif(anim, "my_name.gif", fps=15).

WaterWaves1D.interpolate — Method

interpolate(mesh,vector,x;fast)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh, on collocation points given by x.

If the collocation points x are regularly spaced and the optional keyword argument fast is set to true (default is false), then the algorithm is faster and uses less allocations, but is less precise.

Returns the vector of values on collocation points.

WaterWaves1D.interpolate — Method

interpolate(mesh,vector;n=2^3)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh.

Returns (new_mesh,new_vector) a new uniform mesh with n times as many values, and the vector of values on collocation points.

WaterWaves1D.plot_difference! — Method

plot_difference!( plt, problems ; t,x,interpolation,compression,fast,surface,velocity,fourier,label )

Plots in plt the difference between solutions of initial-value problems at a given time.

Argument

pairs is a either a collection (vector, list, etc.) of elements of type Problem, or of pairs of such elements. If pairs are provided, then solutions of such problems are compared. If a collection of problems are provided, then all possible pairs are compared.

Keyword arguments (all optional)

t is the time. If not provided, then the last computed time is plotted.

if a vector x is provided and if possible, the solution is interpolated to the collocation points x.

if interpolation is provided as an integer, the solution is interpolated on as many collocation points (if true, then the value 2^3 is chosen, default is false).

if fast = false (default is false), then the second problem is interpolated to the collocation points of the first problem.

if compression is provided as an integer m, only one in m points are plotted (if true, then the value 2^3 is chosen, default is false).

surface, velocity and fourier (booleans) determine respectively whether surface deformation, η, tangential velocity, v, and the Fourier coefficients of η (in log-scale) are plotted.

label defines the label(s) if provided (otherwise labels are inferred from the problems).

WaterWaves1D.plot_difference — Method

plot_difference( pairs; t,x, interpolation,compression, surface,velocity,fourier, label )

Same as plot_difference! but generates and returns the plot.

WaterWaves1D.plot_solution! — Method

plot_solution!( plt; problems; t,x,interpolation,compression,surface,velocity,fourier,label )

Plots in plt the solution of initial-value problems at a given time.

Argument

problems is either an element or a collection (vector, list, etc.) of elements of type Problem.

Keyword arguments (all optional)

t is the time. If not provided, then the last computed time is plotted.

if a vector x is provided and if possible, the solution is interpolated to the collocation points x.

if interpolation is provided as an integer, the solution is interpolated on as many collocation points (if true, then the value 2^3 is chosen, default is false).

if compression is provided as an integer m, only one in m points are plotted (if true, then the value 2^3 is chosen, default is false).

surface, velocity and fourier (booleans) determine respectively whether surface deformation, η, tangential velocity, v, and the Fourier coefficients of η (in log-scale) are plotted.

label defines the label(s) if provided (otherwise labels are inferred from the problems)

WaterWaves1D.plot_solution — Method

plot_solution( problems; t,x, interpolation,compression, surface,velocity,fourier, label )

Same as plot_solution! but generates and returns the plot.

WaterWaves1D.random — Method

random(x;args)

Randomly generate a vector of size x, based on provided (optional arguments) :

L is the typical wavelength (default is L=1),

s is the (real) Sobolev index regularity (default is s=∞, smooth data),

λ is the length of spatial localization (default is λ=∞, no localization),

a is the amplitude of the returned vector (default is a=1).

The vector is generated through randomly chosen Fourier coefficients, multiplied with weigth w=10^(-|k|L/(2π)) if s=∞, or w=1/(1+9(|k|L/(2π))^(s+1/2)) otherwise. If λ≠∞, the function in spatial variables is multiplied by exp(-|x/λ|^2), and in any case normalized to have maximum absolute value 1.

WaterWaves1D.solution — Method

solution(p::Problem;t,x,interpolation)

Gives the solution of a solved initial-value at a given time t.

Arguments

Argument p is of type Problem.

Keyword argument t is optional, the last computed time is returned by default.

Keyword argument x is optional, if provided the solution is interpolated to the collocation vector x.

Keyword argument interpolation is optional, if an integer is provided the solution is interpolated on as many collocation points (if true, then the default value 2^3 is chosen).

Return values

Provides (η,v,x,t) where

η is the surface deformation at collocation points;

v is the tangential velocity (derivative of the trace of the velocity potential) at collocation points;

x is the vector of collocation points;

t the time (first computed time greater or equal to provided t).

WaterWaves1D.solve! — Method

solve!( problems; verbose=true )

Solve (i.e. integrate in time) a collection of initial-value problems.

The argument problems should be a collection (list, array...) of elements of type Problem.

Information are not printed if keyword argument verbose = false (default is true).

WaterWaves1D.solve! — Method

solve!( problem :: Problem; verbose=true )

Solve (i.e. integrate in time) an initial-value problem

The argument problem should be of type Problem. It may be buit, e.g., by Problem(model, initial, param)

Information are not printed if keyword argument verbose = false (default is true).