Models

WaterWaves1D.Boussinesq — Type

Boussinesq(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for abcd-Boussinesq models (with b=d and c=0). See Bona, Chen, and Saut

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

two parameters a (default is -1/3) and b (default is +1/3) which determine the model solved. You need a+2*b=1/3 for validity as a long wave model (without surface tension).

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Boussinesq");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function Boussinesq.f! to be called in explicit time-integration solvers;

a function Boussinesq.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function Boussinesq.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.CnoidalSGN — Type

CnoidalSGN(param; P=1)

Build the initial data associated with CnoidalWaveSerreGreenNaghdi(param; P=1), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

WaterWaves1D.Data — Type

Data( mm :: Matrix )

Data structure to store the solution of an initial-value problem along time.

data=Data(m) is of parametric type and offers

data.U, a 1-element vector with a copy of the matrix m;

(data.datalength,data.datasize)=size(m) where datalength is the number of computed modes, and datasize the number of involved equations, typically 2.

WaterWaves1D.DeepQuadratic — Type

DeepQuadratic(param;dealias,label)

Define an object of type AbstractModel in view of solving the initial-value problem for the quadratic deep-water model proposed by Akers and Milewski and Cheng, Granero-Belinchón, Shkoller and Milewski

Arguments

param is of type NamedTuple and must contain

the dimensionless parameters ϵ (nonlinearity);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

dealias: dealiasing with 1/3 Orlicz rule if true or no dealiasing if false (by default);

label: a label for future references (default is "deep quadratic");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function DeepQuadratic.f! to be called in explicit time-integration solvers;

a function DeepQuadratic.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function DeepQuadratic.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is given by ∂t η = - ∂x v.

WaterWaves1D.DeepQuadratic_fast — Type

DeepQuadratic_fast(param;dealias,label)

Same as DeepQuadratic, but faster.

WaterWaves1D.Euler — Type

Euler(arguments;realdata)

Explicit Euler solver.

Construct an object of type TimeSolver to be used in Problem(model, initial, param; solver::TimeSolver)

Arguments can be either

an object of type AbstractModel;

an Array of size (N,m) where N is the number of collocation points and m the number of data (equations solved);

a Tuple (N,m) as above;

an integer N and an integer m as above (the latter is optional, by default m=2).

a NamedTuple containing a key N and an integer m (the latter is optional, by default m=2).

The keyword argument realdata is optional, and determines whether pre-allocated vectors are real- or complex-valued. By default, they are either determined by the model or the type of the array in case 0. and 1., complex-valued otherwise.

WaterWaves1D.EulerSymp — Type

EulerSymp(arguments;Niter,implicit,realdata)

Symplectic Euler solver. The implicit Euler method is first used on one equation, then the explicit Euler method is used on the second one. The implicit equation is solved via Neumann iteration

Construct an object of type TimeSolver to be used in Problem(model, initial, param; solver::TimeSolver)

Arguments can be either

an object of type AbstractModel;

an Array of size (N,2) where N is the number of collocation points;

an integer N being the number of collocation points;

a NamedTuple containing a key N.

The keyword argument Niter (optional, defaut value = 10) determines the number of steps in the Neumann iteration solver of the implicit step. The keyword argument implicit (optional, defaut value = 1) determines which equation is implicit (must be 1 or 2). The keyword argument realdata is optional, and determines whether pre-allocated vectors are real- or complex-valued. By default, they are either determined by the model or the type of the array in case 0. and 1., complex-valued otherwise.

WaterWaves1D.Euler_naive — Type

Euler_naive()

Runge-Kutta fourth order solver.

A naive version of Euler, without argument since no pre-allocation is performed.

WaterWaves1D.Init — Type

Init(data ; fast, label)

Generate an initial data to be used in the function Problem.

data should contain either

a function η and a function v (in this order)

a Namedtuple with a function η and a function v

a mesh and two vectors representing η(mesh.x) and v(mesh.x) (in this order)

a mesh and a Namedtuple with a vector η and a vector v as above

an array of collocation points and two vectors representing η(x) and v(x) (in this order)

a mesh and a Namedtuple with a vector η and a vector v as above

In the last four cases, an optional keyword argument fast can be set to true, (default is false), in which case the algorithm is faster and uses less allocations, but is less precise.

In the last two cases, the collocation points must be regularly spaced, otherwise an ErrorException is raised.

If the keyword label::String (used to display information to the output stream) is not provided, then it is set to the "user-defined".

WaterWaves1D.IsobeKakinuma — Type

IsobeKakinuma(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the Isobe-Kakinuma model.

Argument

param is of type NamedTuple (or a collection NamedTuples) of and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Isobe-Kakinuma");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function IsobeKakinuma.f! to be called in explicit time-integration solvers;

a function IsobeKakinuma.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function IsobeKakinuma.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function IsobeKakinuma.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,Φ), where

Φ is the Vector of the basis functions ϕi (i∈{0,...,N}).

WaterWaves1D.Matsuno — Type

Matsuno(param;dealias,label)

Define an object of type AbstractModel in view of solving the initial-value problem for the quadratic deep-water model proposed by Matsuno.

Arguments

param is of type NamedTuple and must contain

the dimensionless parameters ϵ (nonlinearity);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

dealias: dealiasing with 1/3 Orlicz rule if true or no dealiasing if false (by default);

label: a label for future references (default is "Matsuno");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function DeepQuadratic.f! to be called in explicit time-integration solvers;

a function DeepQuadratic.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function DeepQuadratic.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is a velocity variable which is not the derivative of the trace of the velocity potential (if not null).

WaterWaves1D.Matsuno_fast — Type

Matsuno_fast(param;dealias,label)

Same as Matsuno, but faster.

WaterWaves1D.Mesh — Type

Mesh(args)

Constructs a mesh of collocation points and associated Fourier modes.

Arguments

Can be either

xmin, xmax, and N; or

L, N (same as above with xmin=-L and xmax=L); or

param :: NamedTuple, a NamedTuple containing N and L or xmin and xmax, then same as above; or

x a vector of regularly spaced collocation points`.

The mesh as N collocation points regularly spaced between xmin (included) and xmax (excluded)

Return values

m=Mesh(args) is of parametric type and offers with

m.N: number of collocation points and Fourier modes;

m.xmin: minimum of the mesh (included in the vector of collocation points);

m.xmax: maximum of the mesh (excluded in the vector of collocation points);

m.dx: distance between two collocation points;

m.x: the vector of collocation points;

m.kmin: minimum of Fourier modes (included in the vector of Fourier modes);

m.kmax: maximum of Fourier modes (included in the vector of Fourier modes);;

m.dk: distance between two Fourier modes;

m.k: the vector of Fourier modes.

WaterWaves1D.NonHydrostatic — Type

NonHydrostatic(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the "Non-hydrostatic" model proposed by Bristeau, Mangeney, Sainte-Marie and Seguin

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "non-hydrostatic");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function NonHydrostatic.f! to be called in explicit time-integration solvers;

a function NonHydrostatic.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function NonHydrostatic.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function NonHydrostatic.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.Problem — Type

Problem( model, initial, param ; solver, label)

Build an initial-value problem which can then be solved (i.e. integrated in time) through solve!( problem )

Arguments

model :: AbstractModel, the system of equation solved.

May be built, e.g., by WaterWaves(param);

initial :: InitialData, the initial data.

May be buit, e.g., by Init(η,v) where η is the surface deformation and v the derivative of the trace of the velocity potential at the surface;

param :: NamedTuple, must contain values for

N, the number of collocation points of the spatial grid

L, the half-length of the spatial grid

T, the final time of integration

dt, the timestep

additionally, it may contain Ns the number of computed data or ns for storing data every ns computation steps (by default, every computed data is stored).

Optional keyword arguments

solver :: TimeSolver, the solver for time integration (default is explicit Runge-Kutta fourth order solver).

May be built, e.g., by RK4(model) or RK4_naive().

label :: String is used in future references (e.g.plot_solution`).

Information are not printed if verbose = false (default is true).

WaterWaves1D.RK4 — Type

RK4(arguments;realdata)

Explicit Runge-Kutta fourth order solver.

Construct an object of type TimeSolver to be used in Problem(model, initial, param; solver::TimeSolver)

Arguments can be either

an object of type AbstractModel;

an Array of size (N,m) where N is the number of collocation points and m the number of data (equations solved);

a Tuple (N,m) as above;

an integer N and an integer m as above (the latter is optional, by default m=2).

a NamedTuple containing a key N and an integer m (the latter is optional, by default m=2).

The keyword argument realdata is optional, and determines whether pre-allocated vectors are real- or complex-valued. By default, they are either determined by the model in case 1., complex-valued otherwise.

WaterWaves1D.RK4_naive — Type

RK4_naive()

Runge-Kutta fourth order solver.

A naive version of RK4, without argument since no pre-allocation is performed.

WaterWaves1D.Random — Type

Random(mesh::Mesh;args)

Randomly generated initial data, based on provided (optional arguments) :

L is the typical wavelength (default is L=1),

s is the (real) Sobolev index regularity (default is s=∞),

λ is the length of spatial localization (default is λ=∞, no localization),

a is the couple of amplitudes of the surface deformation, and velocity (default is a=(1,1)).

Return an initial data init::InitialData, to be used in initial-value problems Problem(model, init, param), with init.η and init.v two samples of random(x;L,s,λ), where x is the set of collocation points generated by Mesh(param).

WaterWaves1D.SerreGreenNaghdi — Type

SerreGreenNaghdi(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the Serre-Green-Naghdi model

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Green-Naghdi");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function SerreGreenNaghdi.f! to be called in explicit time-integration solvers;

a function SerreGreenNaghdi.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function SerreGreenNaghdi.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function SerreGreenNaghdi.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.SolitarySGN — Type

SolitarySGN(param; x₀=0)

Build the initial data associated with SolitaryWaveSerreGreenNaghdi(param; x₀=0), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

SolitarySGN(c; ϵ=1,μ=1,x₀=0,N=2^12)

Build the initial data with velocity c, center x₀, dimensionless parameters ϵ and μ, and number of collocation points N.

WaterWaves1D.SolitaryWB — Type

SolitaryWB(param; kwargs)

Build the initial data associated with SolitaryWaveWhithamBoussinesq(param; kwargs), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

SolitaryWB(c; ϵ=1,μ=1,N=2^12,kwargs)

Build the initial data with velocity c, dimensionless parameters ϵ and μ, and number of collocation points N, and kwargs the other (optional) keyword arguments as above.

WaterWaves1D.SolitaryWGN — Type

SolitaryWGN(param; kwargs)

Build the initial data associated with SolitaryWaveWhithamGreenNaghdi(param; kwargs), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

SolitaryWGN(c; ϵ=1,μ=1,N=2^12,kwargs)

Build the initial data with velocity c, dimensionless parameters ϵ and μ, and number of collocation points N, and kwargs the other (optional) keyword arguments as above.

WaterWaves1D.SolitaryWhitham — Type

SolitaryWhitham(param; kwargs)

Build the initial data associated with SolitaryWaveWhitham(param; kwargs), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

SolitaryWhitham(c; ϵ=1,μ=1,N=2^12,kwargs)

Build the initial data with velocity c, dimensionless parameters ϵ and μ, and number of collocation points N, and kwargs the other (optional) keyword arguments as above.

WaterWaves1D.SquareRootDepth — Type

SquareRootDepth(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the "√D" model proposed by Cotter, Holm and Percival

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "square-root depth");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function SquareRootDepth.f! to be called in explicit time-integration solvers;

a function SquareRootDepth.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function SquareRootDepth.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function SquareRootDepth.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.Times — Type

Times(param; ns, Ns)

Constructs a mesh of times, to be used in initial-value problems (see Problem).

Arguments

param is either

dt,T with dt the timestep and T the final time of comuptation; or

a NamedTuple containing dt and T

Optional keyword arguments

ns : data are stored every ns computations (optional, default = 1).

Ns : Ns data (in addition to the initial datum) are stored (optional, by default `floor( tfin/dt)).

If both Ns and ns are given, Ns overrules ns.

Return values

t=Times(args) is of parametric type and offers

t.Nc: number of computed times (including initial datum);

t.Ns: number of stored times (including initial datum);

t.ns: number of computed times between two stored times;

t.tfin: the final time;

t.dt: the timestep;

t.tc : the vector of computed times;

t.ts: the vector of stored times.

WaterWaves1D.WWn — Type

WWn(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the water waves expansion proposed by Dommermuth and Yue (doi:10.1017/s002211208700288x), West et al. (doi:10.1029/jc092ic11p11803), Craig and Sulem (doi:10.1006/jcph.1993.1164) (see also the account by Choi : Fifth-order nonlinear spectral model for surface gravity waves: From pseudo-spectral to spectral formulations, RIMS Kokyuroku, 2019) with the "rectification" method proposed by Duchêne and Melinand.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

ν: shallow/deep water multiplication factor. By default, ν=1 if μ≦1 and ν=1/√μ otherwise. Set the infinite-layer case if ν=0 (or μ=Inf).

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

n :: Int: the order of the expansion; linear system if 1, quadratic if 2, cubic if 3, quartic if 4 (default and other values yield 2);

δ and m: parameters of the rectifier operator, set as k->min(1,|δ*k|^m) or k->min(1,|δ*k|^m[1]*exp(1-|δ*k|^m[2])) if m is a couple

(by default is δ=0, i.e. no regularization and m=-1. Notice m=-Inf and δ>0 yields a cut-off filter);

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "WWn" with n the order of the expansion);

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WWn.f! to be called in explicit time-integration solvers (also WWn.f1! and WWn.f2! for the symplectic Euler solver);

a function WWn.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WWn.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.WaterWaves — Type

WaterWaves(param; kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the water waves system (via conformal mapping, see Zakharov, Dyachenko and Vasilyev).

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

ν: shallow/deep water multiplication factor (see Lannes, The water waves problem). By default, ν=1 if μ≦1 and ν=1/√μ otherwise.

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

method ∈ {1,2,3}: method used to initialize the conformal mapping, as a fix-point problem F(u)=u

if method == 1, use standard contraction fix-point iteration;

if method == 2, use Newton algorithm with GMRES iterative solver to invert the Jacobian;

if method == 3, use Newton algorithm with direct solver to invert the Jacobian;

tol: (relative) tolerance of the fix-point algorithm (default is 1e-16);

maxiter: the maximal number of iteration in the fix-point algorithm (default is 100);

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "water waves");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WaterWaves.f! to be called in the explicit time-integration solver (also WaterWaves.f1! and WaterWaves.f2! for the symplectic Euler solver);

a function WaterWaves.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WaterWaves.mapfro which from such data matrix returns the Tuple of real vectors (x,η,v), where

x is a vector of collocation points (non-regularly spaced);

η is the surface deformation at points x;

v is the derivative of the trace of the velocity potential at points x.

WaterWaves1D.WhithamBoussinesq — Type

WhithamBoussinesq(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for a Boussinesq-type model with full-dispersion property.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

Boussinesq: if true (default is false), compute the standard Boussinesq system instead (see Boussinesq(param;kwargs));

a parameter α which determines the model solved:

If α = 1 (default), then the model has been introduced in Dinvay, Dutykh and Kalisch;

If α = 1/2, then the model is a quasilinear version;

If α < 1/2, then expect instabilities stemming from ill-posedness of the model.

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Whitham-Boussinesq");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WhithamBoussinesq.f! to be called in explicit time-integration solvers;

a function WhithamBoussinesq.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed.

a function WhithamBoussinesq.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.WhithamGreenNaghdi — Type

WhithamGreenNaghdi(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the fully dispersive Green-Naghdi model proposed by Duchêne, Israwi and Talhouk.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

SGN: if true (default is false), compute the Serre-Green-Naghdi (SGN) instead of Whitham-Green-Naghdi (WGN) system (see SerreGreenNaghdi(param;kwargs));

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Whitham-Green-Naghdi");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WhithamGreenNaghdi.f! to be called in explicit time-integration solvers;

a function WhithamGreenNaghdi.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WhithamGreenNaghdi.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function WhithamGreenNaghdi.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.modifiedMatsuno — Type

modifiedMatsuno(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the modified Matsuno model

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

ν: shallow/deep water multiplication factor. By default, ν=1 if μ≦1 and ν=1/√μ otherwise. Set the infinite-layer case if ν=0 (or μ=Inf).

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "modified Matsuno");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function modifiedMatsuno.f! to be called in explicit time-integration solvers;

a function modifiedMatsuno.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function modifiedMatsuno.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.