WaterWaves1D.jl

WaterWaves1D.BellCurve — Type

BellCurve(param)
param should contain a value theta

\[\eta = 2^{(-|x|^\theta)}\]

\[v = 0\]

WaterWaves1D.Data — Type

Data structure to store the solution of the problem along time

WaterWaves1D.Init — Type

Init(data)
data should contain either
- a function η and a function v (in this order)
- a Namedtuple with a function η and a function v
- a mesh and two vectors Vector{Complex{Float64}} or Vector{Float64} representing η(mesh.x) and v(mesh.x) (in this order)
- a mesh and a Namedtuple with a vector η and a vector v as above

WaterWaves1D.Mesh — Type

`Mesh(args)`

Constructs a mesh of collocation points and associated Fourier modes.

Arguments

Can be either

xmin , xmax , N; or

L , N (same as above with xmin=-L and xmax=L); or

param :: NamedTuple, param contains N and L, then same as above.

The mesh as N collocation points regularly spaced between xmin (included) and xmax (excluded)

Return values

m=Mesh(args) is of parametric type and offers with

m.N: number of collocation points and Fourier modes;

m.xmin: minimum of the mesh;

m.xmax: maximum of the mesh;

m.dx: distance between two collocation points;

m.x: the vector of collocation points;

m.kmin: minimum of Fourier modes;

m.kmax: maximum of Fourier modes;

m.dk: distance between two Fourier modes;

m.k: the vector of Fourier modes.

WaterWaves1D.PseudoSpectral — Type

PseudoSpectral(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the modified water waves expansion proposed by West et al., Craig-Sulem, etc.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

ν: shallow/deep water multiplication factor. By default, ν=1 if μ≦1 and ν=1/√μ otherwise. Set the infinite-layer case if ν=0 (or μ=Inf).

order :: Int: the order of the expansion; linear system if 1, quadratic if 2, cubic if 3, quartic if 4 (default and other values yield 2);

δ: parameter of smooth regularization operator (default is 0, i.e. no regularization);

reg: order of the smooth regularization operator (default is 1);

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve! and in particular

a function PseudoSpectral.f! to be called in the time-integration solver;

a function PseudoSpectral.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function PseudoSpectral.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.Times — Type

`Times(dt, tfin; ns, Ns)`

Constructs a mesh of times.

Arguments

dt : timestep;

tfin: final time;

ns : data are stored every ns computations (optional, default = 1).

Ns : Ns data (in addition to the initial datum) are stored (optional, by default `floor( tfin/dt)).

If both Ns and ns are given, Ns overrules ns.

Return values

t=Times(args) is of parametric type and offers

t.Nc: number of computed times;

t.Ns: number of stored times;

t.ns: number of computed times between two stored times;

t.tfin: the final time;

t.dt: the timestep;

t.tc : the vector of computed times;

t.ts: the vector of stored times.

WaterWaves1D.WhithamGreenNaghdi — Type

WhithamGreenNaghdi(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the modified Green-Naghdi model proposed by V. Duchêne, S. Israwi and R. Talhouk.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Keywords

SGN: if true computes the Serre-Green-Naghdi (SGN) instead of Whitham-Green-Naghdi (WGN) system (default is false);

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve! and in particular

a function WhithamGreenNaghdi.f! to be called in the time-integration solver;

a function WhithamGreenNaghdi.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WhithamGreenNaghdi.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function WhithamGreenNaghdi.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

- `u` corresponds to the layer-averaged velocity.

WaterWaves1D.interpolate — Method

interpolate(mesh,vector,x)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh, on collocation points given by x.

Returns new_vector the vector of values on collocation points.

WaterWaves1D.interpolate — Method

interpolate(mesh,vector;n=2^4)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh.

Returns (new_mesh,new_vector) a new uniform mesh with n times as many values, and the vector of values on collocation points.

WaterWaves1D.loadpb — Method

`loadpb(name::String)`

Loads the contents of the file name.jld as a problem of type :Problem.

WaterWaves1D.solve! — Method

solve!( problems; verbose=true )

Solves (i.e. integrates in time) a collection of initial-value problems.

The argument problems should be a collection (list, array...) of elements of type Problem.