WaterWaves1D.jl

WaterWaves1D.Data — Type

Data structure to store the solution of the problem along time

WaterWaves1D.Init — Type

Init(data)
data should contain either
- a function η and a function v (in this order)
- a Namedtuple with a function η and a function v
- a mesh and two vectors Vector{Complex{Float64}} or Vector{Float64} representing η(mesh.x) and v(mesh.x) (in this order)
- a mesh and a Namedtuple with a vector η and a vector v as above

WaterWaves1D.Mesh — Type

`Mesh(args)`

Constructs a mesh of collocation points and associated Fourier modes.

Arguments

Can be either

xmin , xmax , N; or

L , N (same as above with xmin=-L and xmax=L); or

param :: NamedTuple, param contains N and L, then same as above; or

x a vector of regularly spaced collocation points`.

The mesh as N collocation points regularly spaced between xmin (included) and xmax (excluded)

Return values

m=Mesh(args) is of parametric type and offers with

m.N: number of collocation points and Fourier modes;

m.xmin: minimum of the mesh;

m.xmax: maximum of the mesh;

m.dx: distance between two collocation points;

m.x: the vector of collocation points;

m.kmin: minimum of Fourier modes;

m.kmax: maximum of Fourier modes;

m.dk: distance between two Fourier modes;

m.k: the vector of Fourier modes.

WaterWaves1D.Times — Type

`Times(dt, tfin; ns, Ns)`

Constructs a mesh of times.

Arguments

dt : timestep;

tfin: final time;

ns : data are stored every ns computations (optional, default = 1).

Ns : Ns data (in addition to the initial datum) are stored (optional, by default `floor( tfin/dt)).

If both Ns and ns are given, Ns overrules ns.

Return values

t=Times(args) is of parametric type and offers

t.Nc: number of computed times;

t.Ns: number of stored times;

t.ns: number of computed times between two stored times;

t.tfin: the final time;

t.dt: the timestep;

t.tc : the vector of computed times;

t.ts: the vector of stored times.

WaterWaves1D.create_animation — Method

create_animation( problems; name::String, kwargs... )

Create an animation showing the evolution of initial-value problems.

Argument problems is either an element or a collection (vector, list, etc.) of elements of type Problem.

The animation is saved as name.gif if name is provided.

Other keyword arguments are as follows

xlims allows to specifies the x axis limits for the surface deformation. If nothing is provided (default), then the full numerical basin is represented.

ylims allows to specifies the y axis limits for the surface deformation. If nothing is provided (default), then the limits are determined from the initial data. If anything but a Tuple is provided, the axis limits evolve with the solution.

vlims and flims are as above, but for the velocity and Fourier coefficients plots.

Nframes gives the (maximal) number of frames in the animation.

other arguments of plot_solution!

Return anim, an animation, which can then generate (for instance) a gif through gif(anim, "my_name.gif", fps=15).

WaterWaves1D.interpolate — Method

interpolate(mesh,vector,x)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh, on collocation points given by x.

Returns new_vector the vector of values on collocation points.

WaterWaves1D.interpolate — Method

interpolate(mesh,vector;n=2^3)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh.

Returns (new_mesh,new_vector) a new uniform mesh with n times as many values, and the vector of values on collocation points.

WaterWaves1D.plot_solution! — Method

plot_solution!( plt; problems; t,x,interpolation,compression,surface,velocity,fourier,label )

Plots in plt the solution of initial-value problems at a given time.

Argument

problems is either an element or a collection (vector, list, etc.) of elements of type Problem.

Keywords

t is the time. If not provided, then the last computed time is plotted.

if a vector x is provided and if possible, the solution is interpolated to the collocation points x.

if interpolation is provided as an integer, the solution is interpolated on as many collocation points (if true, then the value 2^3 is chosen, default is false).

if compression is provided as an integer m, only one in m points are plotted (if true, then the value 2^3 is chosen, default is false).

surface, velocity and fourier (booleans) determine respectively whether surface deformation, η, tangential velocity, v, and the Fourier coefficients of η (in log-scale) are plotted.

WaterWaves1D.plot_solution — Method

plot_solution( problems; t,x, interpolation,compression, surface,velocity,fourier, label )

Same as plot_solution! but generates and returns the plot.

WaterWaves1D.solve! — Method

solve!( problems; verbose=true )

Solves (i.e. integrates in time) a collection of initial-value problems.

The argument problems should be a collection (list, array...) of elements of type Problem.