ShallowWaterModels.jl

ShallowWaterModels.Data — Type

Data structure to store the solution of the problem along time

ShallowWaterModels.Init — Type

Init(data)
data should contain either
- a function η and a function v (in this order)
- a Namedtuple with a function η and a function v
- a mesh and two vectors Vector{Complex{Float64}} or Vector{Float64} representing η(mesh.x) and v(mesh.x) (in this order)
- a mesh and a Namedtuple with a vector η and a vector v as above

ShallowWaterModels.Mesh — Type

`Mesh(args)`

Constructs a mesh of collocation points and associated Fourier modes.

Arguments

Can be either

xmin , xmax , N; or

L , N (same as above with xmin=-L and xmax=L); or

param :: NamedTuple, param contains N and L, then same as above.

The mesh as N collocation points regularly spaced between xmin (included) and xmax (excluded)

Return values

m=Mesh(args) is of parametric type and offers with

m.N: number of collocation points and Fourier modes;

m.xmin: minimum of the mesh;

m.xmax: maximum of the mesh;

m.dx: distance between two collocation points;

m.x: the vector of collocation points;

m.kmin: minimum of Fourier modes;

m.kmax: maximum of Fourier modes;

m.dk: distance between two Fourier modes;

m.k: the vector of Fourier modes.

ShallowWaterModels.Times — Type

`Times(dt, tfin; ns, Ns)`

Constructs a mesh of times.

Arguments

dt : timestep;

tfin: final time;

ns : data are stored every ns computations (optional, default = 1).

Ns : Ns data (in addition to the initial datum) are stored (optional, by default `floor( tfin/dt)).

If both Ns and ns are given, Ns overrules ns.

Return values

t=Times(args) is of parametric type and offers

t.Nc: number of computed times;

t.Ns: number of stored times;

t.ns: number of computed times between two stored times;

t.tfin: the final time;

t.dt: the timestep;

t.tc : the vector of computed times;

t.ts: the vector of stored times.

ShallowWaterModels.interpolate — Method

interpolate(mesh,vector,x)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh, on collocation points given by x.

Returns new_vector the vector of values on collocation points.

ShallowWaterModels.interpolate — Method

interpolate(mesh,vector;n=2^4)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh.

Returns (new_mesh,new_vector) a new uniform mesh with n times as many values, and the vector of values on collocation points.

ShallowWaterModels.solve! — Method

solve!( problems; verbose=true )

Solves (i.e. integrates in time) a collection of initial-value problems.

The argument problems should be a collection (list, array...) of elements of type Problem.