ShallowWaterModels.jl

ShallowWaterModels.BellCurve — Type

BellCurve(param)
param should contain a value theta

\[\eta = 2^{(-|x|^\theta)}\]

\[v = 0\]

ShallowWaterModels.Boussinesq — Type

Boussinesq(params)
params must contain two parameters a and b.
This computes the abcd Boussinesq model with d=b and c=0. You need a+2*b=1/3 for validity as a long wave model.

ShallowWaterModels.CGBSW — Type

CGBSW( params )

ShallowWaterModels.CGBSW_naive — Type

CGBSW_naive( params )

ShallowWaterModels.Data — Type

Data structure to store the solution of the problem along time

ShallowWaterModels.HighFreq — Type

HighFreq(param)
param should contain s (regularity index) and freq (frequencies)

ShallowWaterModels.Init — Type

Init(data)
data should contain either
- a function η and a function v (in this order)
- a Namedtuple with a function η and a function v
- a mesh and two vectors Vector{Complex{Float64}} or Vector{Float64} representing η(mesh.x) and v(mesh.x) (in this order)
- a mesh and a Namedtuple with a vector η and a vector v as above

ShallowWaterModels.Matsuno_mod_naive — Type

Modified Matsuno models with a naive step function
Matsuno_mod_naive(params)

ShallowWaterModels.Matsuno_naive — Type

Matsuno(params)

ShallowWaterModels.Mesh — Type

`Mesh(args)`

Constructs a mesh of collocation points and associated Fourier modes.

Arguments

Can be either

xmin :: Float64, xmax :: Float64, N :: Int64;

L :: Float64, N :: Int64, same as above with xmin=-L and xmax=L;

param :: NamedTuple, param contains N and L, then same as above.

The mesh as N collocation points regularly spaced between xmin (included) and xmax (excluded)

Return values

m=Mesh(args) is of parametric type and offers with

m.N: number of collocation points and Fourier modes;

m.xmin: minimum of the mesh;

m.xmax: maximum of the mesh;

m.dx: distance between two collocation points;

m.x: the vector of collocation points;

m.kmin: minimum of Fourier modes;

m.kmax: maximum of Fourier modes;

m.dk: distance between two Fourier modes;

m.k: the vector of Fourier modes.

ShallowWaterModels.PseudoSpectral — Type

PseudoSpectral(param;kwargs)

Defines an object of type AbstractModel in view of solving the initial-value problem for the modified water waves expansion proposed by West et al., Craig-Sulem, etc.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

E.g. param = ( μ = 0.1, ϵ = 1, N = 2^9, L = 10*π)

Keywords

order :: Int: the order of the expansion; linear system if 1, quadratic if 2, cubic if 3, quartic if 4 (default and other values yield 2);

lowpass :: Real: applies a low-pass filter on frequencies above lowpass (default is 0, i.e. no low-pass filter);

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing).

Return values

This generates

a function PseudoSpectral to be called in the time-integration solver;

a function mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed.

a function mapfro which from such raw data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is a tangential surface velocity (the derivative of the trace of the velocity potential at the surface).

ShallowWaterModels.RK4_naive — Type

`RK4_naive()`

Runge-Kutta fourth order solver. Constructs an object of type TimeSolver to be used in `Problem(model, initial, param; solver::TimeSolver)

ShallowWaterModels.Random — Type

Random(param)

param should contain s (regularity index)

ShallowWaterModels.Times — Type

`Times(dt, tfin; ns = 1)`

Constructs a mesh of times.

Arguments

dt : timestep;

tfin: final time;

ns : data are stored every nr computations (optional, default = 1).

Return values

t=Times(args) is of parametric type and offers

t.Nc: number of computed times;

t.Ns: number of recorded (stored) times;

t.ns: data are stored every ns computations;

t.tfin: the final time;

t.dt: the timestep;

t.tc : the vector of computed times;

t.ts: the vector of recorded times.

ShallowWaterModels.WaterWaves — Type

WaterWaves(params)
The Water Waves system, through conformal mapping

ShallowWaterModels.WhithamGreenNaghdi — Type

WhithamGreenNaghdi(param;kwargs)

Defines an object of type AbstractModel in view of solving the initial-value problem for the modified Green-Naghdi model proposed by V. Duchêne, S. Israwi and R. Talhouk.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

E.g. param = ( μ = 0.1, ϵ = 1, N = 2^9, L = 10*π)

Keywords

SGN :: Bool: if true computes the Serre-Green-Naghdi (SGN) instead of Whitham-Green-Naghdi (WGN) system (default is false);

iterative :: Bool: solves the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond :: Bool: Preconditioner of GMRES is based on WGN if true, SGN otherwise (default is true);

gtol :: Real: relative tolerance of the GMRES algorithm (default is 1e-14);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing).

Return values

This generates

a function WhithamGreenNaghdi to be called in the time-integration solver;

a function mapto which from (η,v) of type InitialData provides the data matrix on which computations are to be executed.

a function mapfro which from such data matrix returns the Tuple of real vectors (η,v,u), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

u corresponds to the layer-averaged velocity.

ShallowWaterModels.WhithamGreenNaghdiKlein — Type

WhithamGreenNaghdi(params)
The modified Green-Naghdi model proposed by V. Duchêne, S. Israwi and R. Talhouk.

ShallowWaterModels.WhithamGreenNaghdiSym — Type

WhithamGreenNaghdiSym(params)
The modified Green-Naghdi model proposed by V. Duchêne, S. Israwi and R. Talhouk.

ShallowWaterModels.fdBoussinesq — Type

fdBoussinesq_1(params)
params must contain a parameter α.
If α = 1, then the model has been introduced and studied by E. Dinvay and collaborators.
If α = 1/2, then the model is a quasilinear version.
If α < 1/2, then expect instabilities stemming from ill-posedness of the model.

ShallowWaterModels.SolitaryWaveWhithamBoussinesq — Method

`SolitaryWaveWhithamBoussinesq(mesh, param, guess; kwargs...)`

Computes the Whitham-Boussinesq solitary wave with prescribed velocity.

Arguments

mesh :: Mesh: parameters of the numerical grid, e.g constructed through Mesh(L,N);

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ;

guess :: Vector{Float64}: initial guess for the surface deformation.

Keywords

model :: Real: determines the model used (typically 1 or 1/2, default is 1);

iterative :: Bool: inverts Jacobian through GMRES if true, LU decomposition if false;

verbose :: Bool: prints numerical errors at each step if true;

max_iter :: Int: maximum number of iterations of the Newton algorithm;

tol :: Real: general tolerance (default is 1e-10);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

gtol :: Real: relative tolerance of the GMRES algorithm;

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

q :: Real: Newton algorithm modified with

u_{n+1}=q*u_{n+1}+(1-q)*u_n (default is 1);

α :: Real: adds α times spectral projection onto the Kernel to the Jacobian.

Return values

(η,u) :: Tuple{Vector{Float64},Vector{Float64}} with

η: surface deformation;

u: velocity.

ShallowWaterModels.SolitaryWaveWhithamGreenNaghdi — Method

`SolitaryWaveWhithamGreenNaghdi(mesh, param, guess; kwargs...)`

Computes the Whitham-Green-Naghdi solitary wave with prescribed velocity.

Arguments

mesh :: Mesh: parameters of the numerical grid, e.g constructed through Mesh(L,N);

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ;

guess :: Vector{Float64}: initial guess for the surface deformation.

Keywords

method :: Int: equation used (between 1 and 4);

iterative :: Bool: inverts Jacobian through GMRES if true, LU decomposition if false;

verbose :: Bool: prints numerical errors at each step if true;

max_iter :: Int: maximum number of iterations of the Newton algorithm;

tol :: Real: general tolerance (default is 1e-10);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

gtol :: Real: relative tolerance of the GMRES algorithm;

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

q :: Real: Newton algorithm modified with

u_{n+1}=q*u_{n+1}+(1-q)*u_n (default is 1);

α :: Real: adds α times spectral projection onto the Kernel to the Jacobian;

SGN :: Bool: if true computes the Serre-Green-Naghdi (instead of Whitham-Green-Naghdi) solitary wave.

Return values

(η,u) :: Tuple{Vector{Float64},Vector{Float64}} with

η: surface deformation;

u: velocity.

ShallowWaterModels.mapfro — Method

mapfro(Boussinesq, data)

ShallowWaterModels.mapfro — Method

mapfro(CGBSW, data)

ShallowWaterModels.mapfro — Method

mapfro(CGBSW_naive, data)

ShallowWaterModels.mapfro — Method

mapfro(Matsuno, data)

ShallowWaterModels.mapfro — Method

mapfro(Matsuno, data)

ShallowWaterModels.mapfro — Method

mapfro(Matsuno, data)

ShallowWaterModels.mapfro — Method

mapfro(PseudoSpectral, data)

ShallowWaterModels.mapfro — Method

mapfro(WaterWaves, data)
Reconstructs physical variables from conformal variables

ShallowWaterModels.mapfro — Method

mapfro(WhithamGreenNaghdi, data)

data is of type Array{Complex{Float64},2}, e.g. last(p.data.U) where p is of type Problem.

Returns (η,v,u), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

u corresponds to the layer-averaged velocity.

Inverse Fourier transform and real part, plus solving the elliptic problem for u.

See documentation of WhithamGreenNaghdi for more details.

ShallowWaterModels.mapfro — Method

mapfro(WhithamGreenNaghdi, data)

ShallowWaterModels.mapfro — Method

mapfro(WhithamGreenNaghdiSym, data)

ShallowWaterModels.mapfro — Method

mapfro(fdBoussinesq, data)

ShallowWaterModels.mapto — Method

mapto(Boussinesq, data)

ShallowWaterModels.mapto — Method

mapto(CGBSW, data)
the velocity should be zero

ShallowWaterModels.mapto — Method

mapto(CGBSW_naive, data)
the velocity should be zero

ShallowWaterModels.mapto — Method

mapto(Matsuno, data)
the velocity should be zero

ShallowWaterModels.mapto — Method

mapto(Matsuno, data)
the velocity should be zero

ShallowWaterModels.mapto — Method

mapto(Matsuno, data)
the velocity should be zero

ShallowWaterModels.mapto — Method

mapto(PseudoSpectral, data)

ShallowWaterModels.mapto — Method

mapto(WaterWaves, data)
Constructs conformal variables in the flat strip

ShallowWaterModels.mapto — Method

mapto(WhithamGreenNaghdi, data)

data is of type InitialData, maybe constructed by Init(...).

Performs a discrete Fourier transform with, possibly, dealiasing and Krasny filter.

See documentation of WhithamGreenNaghdi for more details.

ShallowWaterModels.mapto — Method

mapto(WhithamGreenNaghdiKlein, data)

ShallowWaterModels.mapto — Method

mapto(WhithamGreenNaghdiSym, data)

ShallowWaterModels.mapto — Method

mapto(fdBoussinesq, data)