WaterWaves1D.jl

WaterWaves1D.Boussinesq — Type

Boussinesq(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for abcd-Boussinesq models (with b=d and c=0). See Bona, Chen, and Saut

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

two parameters a (default is -1/3) and b (default is +1/3) which determine the model solved. You need a+2*b=1/3 for validity as a long wave model (without surface tension).

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Boussinesq");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function Boussinesq.f! to be called in explicit time-integration solvers;

a function Boussinesq.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function Boussinesq.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.Data — Type

Data( mm :: Matrix )

Data structure to store the solution of an initial-value problem along time.

data=Data(m) is of parametric type and offers

data.U, a 1-element vector with a copy of the matrix m;

(data.datalength,data.datasize)=size(m) where datalength is the number of computed modes, and datasize the number of involved equations, typically 2.

WaterWaves1D.DeepQuadratic — Type

DeepQuadratic(param;dealias,label,verbose)

Define an object of type AbstractModel in view of solving the initial-value problem for the quadratic deep-water model proposed by Akers and Milewski and Cheng, Granero-Belinchón, Shkoller and Milewski

Arguments

param is of type NamedTuple and must contain

the dimensionless parameters ϵ (nonlinearity);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

dealias: dealiasing with 1/3 Orlicz rule if true or no dealiasing if false (by default);

label: a label for future references (default is "deep quadratic");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function DeepQuadratic.f! to be called in explicit time-integration solvers;

a function DeepQuadratic.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function DeepQuadratic.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is given by ∂t η = - ∂x v.

WaterWaves1D.DeepQuadratic_fast — Type

DeepQuadratic_fast(param;dealias,label,verbose)

Same as DeepQuadratic, but faster.

WaterWaves1D.Euler — Type

Euler(arguments;realdata)

Explicit Euler solver.

Construct an object of type TimeSolver to be used in Problem(model, initial, param; solver::TimeSolver)

Arguments can be either

an object of type AbstractModel;

an Array of size (N,m) where N is the number of collocation points and m the number of data (equations solved);

a Tuple (N,m) as above;

an integer N and an integer m as above (the latter is optional, by default m=2).

a NamedTuple containing a key N and an integer m (the latter is optional, by default m=2).

The keyword argument realdata is optional, and determines whether pre-allocated vectors are real- or complex-valued. By default, they are either determined by the model or the type of the array in case 0. and 1., complex-valued otherwise.

WaterWaves1D.EulerSymp — Type

EulerSymp(arguments;Niter,implicit,realdata)

Symplectic Euler solver. The implicit Euler method is first used on one equation, then the explicit Euler method is used on the second one. The implicit equation is solved via Neumann iteration

Construct an object of type TimeSolver to be used in Problem(model, initial, param; solver::TimeSolver)

Arguments can be either

an object of type AbstractModel;

an Array of size (N,2) where N is the number of collocation points;

an integer N being the number of collocation points;

a NamedTuple containing a key N.

The keyword argument Niter (optional, defaut value = 10) determines the number of steps in the Neumann iteration solver of the implicit step. The keyword argument implicit (optional, defaut value = 1) determines which equation is implicit (must be 1 or 2). The keyword argument realdata is optional, and determines whether pre-allocated vectors are real- or complex-valued. By default, they are either determined by the model or the type of the array in case 0. and 1., complex-valued otherwise.

WaterWaves1D.Euler_naive — Type

Euler_naive()

Runge-Kutta fourth order solver.

A naive version of Euler, without argument since no pre-allocation is performed.

WaterWaves1D.Init — Type

Init(data ; fast=false)

Generate an initial data to be used in the function Problem.

data should contain either

a function η and a function v (in this order)

a Namedtuple with a function η and a function v

a mesh and two vectors representing η(mesh.x) and v(mesh.x) (in this order)

a mesh and a Namedtuple with a vector η and a vector v as above

an array of collocation points and two vectors representing η(x) and v(x) (in this order)

a mesh and a Namedtuple with a vector η and a vector v as above

In the last four cases, an optional keyword argument fast can be set to true, (default is false), in which case the algorithm is faster and uses less allocations, but is less precise.

In the last two cases, the collocation points must be regularly spaced, otherwise an ErrorException is raised.

WaterWaves1D.IsobeKakinuma — Type

IsobeKakinuma(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the Isobe-Kakinuma model.

Argument

param is of type NamedTuple (or a collection NamedTuples) of and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Isobe-Kakinuma");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function IsobeKakinuma.f! to be called in explicit time-integration solvers;

a function IsobeKakinuma.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function IsobeKakinuma.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function IsobeKakinuma.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,Φ), where

Φ is the Vector of the basis functions ϕi (i∈{0,...,N}).

WaterWaves1D.Matsuno_fast — Type

Matsuno_fast(param;dealias,label,verbose)

Same as Matsuno, but faster.

WaterWaves1D.Mesh — Type

Mesh(args)

Constructs a mesh of collocation points and associated Fourier modes.

Arguments

Can be either

xmin, xmax, and N; or

L, N (same as above with xmin=-L and xmax=L); or

param :: NamedTuple, a NamedTuple containing N and L or xmin and xmax, then same as above; or

x a vector of regularly spaced collocation points`.

The mesh as N collocation points regularly spaced between xmin (included) and xmax (excluded)

Return values

m=Mesh(args) is of parametric type and offers with

m.N: number of collocation points and Fourier modes;

m.xmin: minimum of the mesh (included in the vector of collocation points);

m.xmax: maximum of the mesh (excluded in the vector of collocation points);

m.dx: distance between two collocation points;

m.x: the vector of collocation points;

m.kmin: minimum of Fourier modes (included in the vector of Fourier modes);

m.kmax: maximum of Fourier modes (included in the vector of Fourier modes);;

m.dk: distance between two Fourier modes;

m.k: the vector of Fourier modes.

WaterWaves1D.NonHydrostatic — Type

NonHydrostatic(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the "Non-hydrostatic" model proposed by Bristeau, Mangeney, Sainte-Marie and Seguin

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "non-hydrostatic");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function NonHydrostatic.f! to be called in explicit time-integration solvers;

a function NonHydrostatic.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function NonHydrostatic.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function NonHydrostatic.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.RK4_naive — Type

RK4_naive()

Runge-Kutta fourth order solver.

A naive version of RK4, without argument since no pre-allocation is performed.

WaterWaves1D.Random — Type

Random(;args)

Randomly generated initial data, based on provided (optional arguments) :

L is the typical wavelength (default is L=1)

s is the (real) Sobolev index regularity (default is s=∞)

λ is the length of spatial localization (default is none)

The initial data (η,v) are generated through randomly chosen Fourier coefficients, multiplied with weigth w=10^(-|k|L/(2π)) if s=∞, or w=1/(1+9(|k|L/(2π))^(s+1/2)) otherwise. If λ is provided, the function in spatial variables is multiplied by exp(-|x/λ|^2), and in any case normalized to have maximum absolute value 1.

WaterWaves1D.SerreGreenNaghdi — Type

SerreGreenNaghdi(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the Serre-Green-Naghdi model

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Green-Naghdi");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function SerreGreenNaghdi.f! to be called in explicit time-integration solvers;

a function SerreGreenNaghdi.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function SerreGreenNaghdi.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function SerreGreenNaghdi.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.SquareRootDepth — Type

SquareRootDepth(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the "√D" model proposed by Cotter, Holm and Percival

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "square-root depth");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function SquareRootDepth.f! to be called in explicit time-integration solvers;

a function SquareRootDepth.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function SquareRootDepth.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function SquareRootDepth.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.Times — Type

Times(param; ns, Ns)

Constructs a mesh of times, to be used in initial-value problems (see Problem).

Arguments

param is either

dt,T with dt the timestep and T the final time of comuptation; or

a NamedTuple containing dt and T

Optional keyword arguments

ns : data are stored every ns computations (optional, default = 1).

Ns : Ns data (in addition to the initial datum) are stored (optional, by default `floor( tfin/dt)).

If both Ns and ns are given, Ns overrules ns.

Return values

t=Times(args) is of parametric type and offers

t.Nc: number of computed times (including initial datum);

t.Ns: number of stored times (including initial datum);

t.ns: number of computed times between two stored times;

t.tfin: the final time;

t.dt: the timestep;

t.tc : the vector of computed times;

t.ts: the vector of stored times.

WaterWaves1D.WWn — Type

WWn(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the water waves expansion proposed by Dommermuth and Yue (doi:10.1017/s002211208700288x), West et al. (doi:10.1029/jc092ic11p11803), Craig and Sulem (doi:10.1006/jcph.1993.1164) (see also the account by Choi : Fifth-order nonlinear spectral model for surface gravity waves: From pseudo-spectral to spectral formulations, RIMS Kokyuroku, 2019) with the "rectification" method proposed by Duchêne and Melinand.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

ν: shallow/deep water multiplication factor. By default, ν=1 if μ≦1 and ν=1/√μ otherwise. Set the infinite-layer case if ν=0 (or μ=Inf).

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

n :: Int: the order of the expansion; linear system if 1, quadratic if 2, cubic if 3, quartic if 4 (default and other values yield 2);

δ and m: parameters of the rectifier operator, set as k->min(1,|δ*k|^m) or k->min(1,|δ*k|^m[1]*exp(1-|δ*k|^m[2])) if m is a couple

(by default is δ=0, i.e. no regularization and m=-1. Notice m=-Inf and δ>0 yields a cut-off filter);

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "WWn" with n the order of the expansion);

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WWn.f! to be called in explicit time-integration solvers (also WWn.f1! and WWn.f2! for the symplectic Euler solver);

a function WWn.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WWn.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.WaterWaves — Type

WaterWaves(param; kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the water waves system (via conformal mapping, see Zakharov, Dyachenko and Vasilyev).

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

ν: shallow/deep water multiplication factor (see Lannes, The water waves problem). By default, ν=1 if μ≦1 and ν=1/√μ otherwise.

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

method ∈ {1,2,3}: method used to initialize the conformal mapping, as a fix-point problem F(u)=u

if method == 1, use standard contraction fix-point iteration;

if method == 2, use Newton algorithm with GMRES iterative solver to invert the Jacobian;

if method == 3, use Newton algorithm with direct solver to invert the Jacobian;

tol: (relative) tolerance of the fix-point algorithm (default is 1e-16);

maxiter: the maximal number of iteration in the fix-point algorithm (default is 100);

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "water waves");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WaterWaves.f! to be called in the explicit time-integration solver (also WaterWaves.f1! and WaterWaves.f2! for the symplectic Euler solver);

a function WaterWaves.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WaterWaves.mapfro which from such data matrix returns the Tuple of real vectors (x,η,v), where

x is a vector of collocation points (non-regularly spaced);

η is the surface deformation at points x;

v is the derivative of the trace of the velocity potential at points x.

WaterWaves1D.WhithamBoussinesq — Type

WhithamBoussinesq(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for a Boussinesq-type model with full-dispersion property.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

Boussinesq: if true (default is false), compute the standard Boussinesq system instead (see Boussinesq(param;kwargs));

a parameter α which determines the model solved:

If α = 1 (default), then the model has been introduced in Dinvay, Dutykh and Kalisch;

If α = 1/2, then the model is a quasilinear version;

If α < 1/2, then expect instabilities stemming from ill-posedness of the model.

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Whitham-Boussinesq");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WhithamBoussinesq.f! to be called in explicit time-integration solvers;

a function WhithamBoussinesq.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed.

a function WhithamBoussinesq.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.WhithamGreenNaghdi — Type

WhithamGreenNaghdi(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the fully dispersive Green-Naghdi model proposed by Duchêne, Israwi and Talhouk.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

SGN: if true (default is false), compute the Serre-Green-Naghdi (SGN) instead of Whitham-Green-Naghdi (WGN) system (see SerreGreenNaghdi(param;kwargs));

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Whitham-Green-Naghdi");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WhithamGreenNaghdi.f! to be called in explicit time-integration solvers;

a function WhithamGreenNaghdi.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WhithamGreenNaghdi.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function WhithamGreenNaghdi.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.modifiedMatsuno — Type

modifiedMatsuno(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the modified Matsuno model

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

ν: shallow/deep water multiplication factor. By default, ν=1 if μ≦1 and ν=1/√μ otherwise. Set the infinite-layer case if ν=0 (or μ=Inf).

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "modified Matsuno");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function modifiedMatsuno.f! to be called in explicit time-integration solvers;

a function modifiedMatsuno.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function modifiedMatsuno.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.CnoidalWaveSerreGreenNaghdi — Method

CnoidalWaveSerreGreenNaghdi(param; P)

Compute the Serre-Green-Naghdi cnoidal wave with prescribed h₀<h₁<h₂. h_1 is the minimum, h_2 is the maximum of the wave. As h₀ -> h₁, the cnoidal wave converges towards the solitary wave. See for instance Gavrilyuk, Nkonga, Shyue and Truskinovsky, doi:10.1088/1361-6544/ab95ac

Arguments

param :: NamedTuple: parameters of the problem containing h₀<h₁<h₂ and dimensionless parameters ϵ and μ, and number of collocation points N.

P :: Int: (keyword, optional, default = 2) the number of periods of the cnoidal wave in the constructed mesh.

Return values

(η,u,v,mesh,param) with

η :: Vector{Float64}: surface deformation;

u :: Vector{Float64}: layer-averaged velocity;

v :: Vector{Float64}: tangential velocity;

mesh :: Mesh: mesh collocation points;

param :: NamedTuple: useful parameters

WaterWaves1D.SolitaryWaveSerreGreenNaghdi — Method

SolitaryWaveSerreGreenNaghdi(param; kwargs...)

Compute the Serre-Green-Naghdi solitary wave with prescribed velocity.

Arguments

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ, and mesh size L and number of collocation points N;

x₀ :: Real: (keyword, optional, default = 0) center of solitary wave.

Return values

(η,u,v) with

η :: Vector{Float64}: surface deformation;

u :: Vector{Float64}: layer-averaged velocity;

v :: Vector{Float64}: tangential velocity;

mesh :: Mesh: mesh collocation points.

WaterWaves1D.SolitaryWaveWhitham — Method

`SolitaryWaveWhitham(param; kwargs...)`

Computes the Whitham solitary wave with prescribed velocity.

Argument

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ, and mesh size L and number of collocation points N;

Keywords (optional)

guess :: Vector{Real}: initial guess for the surface deformation (if not provided, the exact formula for KdV is used);

x₀ :: Real: center of solitary wave (if guess is not provided);

iterative :: Bool: inverts Jacobian through GMRES if true, LU decomposition if false;

verbose :: Bool: prints numerical errors at each step if true;

max_iter :: Int: maximum number of iterations of the Newton algorithm;

tol :: Real: general tolerance (default is 1e-10);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

gtol :: Real: relative tolerance of the GMRES algorithm;

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

q :: Real: Newton algorithm modified with

u_{n+1}=q*u_{n+1}+(1-q)*u_n (default is 1);

α :: Real: adds α times spectral projection onto the Kernel to the Jacobian;

KdV :: Bool: if true computes the KdV (instead of Whitham) solitary wave.

Return values

u :: Vector{Float64} the solution

WaterWaves1D.SolitaryWaveWhithamBoussinesq — Method

`SolitaryWaveWhithamBoussinesq(param; kwargs...)`

Computes the Whitham-Boussinesq solitary wave with prescribed velocity.

Argument

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ, and mesh size L and number of collocation points N;

Keywords (optional)

guess :: Vector{Real}: initial guess for the surface deformation (if not provided, the exact formula for SGN is used);

x₀ :: Real: center of solitary wave (if guess is not provided);

model :: Real: determines the model used (typically 1 or 1/2, default is 1);

iterative :: Bool: inverts Jacobian through GMRES if true, LU decomposition if false;

verbose :: Bool: prints numerical errors at each step if true;

max_iter :: Int: maximum number of iterations of the Newton algorithm;

tol :: Real: general tolerance (default is 1e-10);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

gtol :: Real: relative tolerance of the GMRES algorithm;

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

q :: Real: Newton algorithm modified with

u_{n+1}=q*u_{n+1}+(1-q)*u_n (default is 1);

α :: Real: adds α times spectral projection onto the Kernel to the Jacobian.

Return values

(η,u) :: Tuple{Vector{Float64},Vector{Float64}} with

η: surface deformation;

u: velocity.

WaterWaves1D.SolitaryWaveWhithamGreenNaghdi — Method

SolitaryWaveWhithamGreenNaghdi(param; kwargs...)

Compute the Whitham-Green-Naghdi solitary wave with prescribed velocity.

Arguments

param :: NamedTuple: parameters of the problem containing velocity c and dimensionless parameters ϵ and μ, and mesh size L and number of collocation points N;

Keywords (optional)

guess :: Vector{Real}: initial guess for the surface deformation (if not provided, the exact formula for SGN is used);

x₀ :: Real: center of solitary wave (if guess is not provided);

SGN :: Bool: if true computes the Serre-Green-Naghdi (instead of Whitham-Green-Naghdi) solitary wave (consider SolitaryWaveSerreGreenNaghdi instead);

method :: Int: equation used (between 1 and 4);

iterative :: Bool: inverts Jacobian through GMRES if true, LU decomposition if false (default is false);

verbose :: Bool: prints numerical errors at each step if true (default is false);

max_iter :: Int: maximum number of iterations of the Newton algorithm (default is 20);

tol :: Real: relative tolerance measured in ℓ∞ norm (default is 1e-10);

ktol :: Real: tolerance of the Krasny filter (default is 0, i.e. no filtering);

gtol :: Real: relative tolerance of the GMRES algorithm (default is 1e-10);

dealias :: Int: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

q :: Real: Newton algorithm modified with u_{n+1}=q*u_{n+1}+(1-q)*u_n (default is 1);

α :: Real: adds α times spectral projection onto the Kernel to the Jacobian (default is 0).

Return values

(η,u,v) with

η :: Vector{Float64}: surface deformation;

u :: Vector{Float64}: layer-averaged velocity;

v :: Vector{Float64}: tangential velocity;

mesh :: Mesh: mesh collocation points.

WaterWaves1D.create_animation — Method

create_animation( problems; name::String, kwargs... )

Create an animation showing the evolution of initial-value problems.

Argument problems is either an element or a collection (vector, list, etc.) of elements of type Problem.

The animation is saved as name.gif if name is provided.

Other keyword arguments are as follows

xlims allows to specifies the x axis limits for the surface deformation. If nothing is provided (default), then the full numerical basin is represented.

ylims allows to specifies the y axis limits for the surface deformation. If nothing is provided (default), then the limits are determined from the initial data. If anything but a Tuple is provided, the axis limits evolve with the solution.

vlims and flims are as above, but for the velocity and Fourier coefficients plots.

Nframes gives the (maximal) number of frames in the animation.

other arguments of plot_solution!

Return anim, an animation, which can then generate (for instance) a gif through gif(anim, "my_name.gif", fps=15).

WaterWaves1D.interpolate — Method

interpolate(mesh,vector,x;fast)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh, on collocation points given by x.

If the collocation points x are regularly spaced and the optional keyword argument fast is set to true (default is false), then the algorithm is faster and uses less allocations, but is less precise.

Returns the vector of values on collocation points.

WaterWaves1D.interpolate — Method

interpolate(mesh,vector;n=2^3)

Interpolate a vector vector defined on a uniform collocation grid defined by mesh.

Returns (new_mesh,new_vector) a new uniform mesh with n times as many values, and the vector of values on collocation points.

WaterWaves1D.plot_difference! — Method

plot_difference!( plt, problems ; t,x,interpolation,compression,fast,surface,velocity,fourier,label )

Plots in plt the difference between solutions of initial-value problems at a given time.

Argument

pairs is a either a collection (vector, list, etc.) of elements of type Problem, or of pairs of such elements. If pairs are provided, then solutions of such problems are compared. If a collection of problems are provided, then all possible pairs are compared.

Keyword arguments (all optional)

t is the time. If not provided, then the last computed time is plotted.

if a vector x is provided and if possible, the solution is interpolated to the collocation points x.

if interpolation is provided as an integer, the solution is interpolated on as many collocation points (if true, then the value 2^3 is chosen, default is false).

if fast = false (default is false), then the second problem is interpolated to the collocation points of the first problem.

if compression is provided as an integer m, only one in m points are plotted (if true, then the value 2^3 is chosen, default is false).

surface, velocity and fourier (booleans) determine respectively whether surface deformation, η, tangential velocity, v, and the Fourier coefficients of η (in log-scale) are plotted.

label defines the label(s) if provided (otherwise labels are inferred from the problems).

WaterWaves1D.plot_difference — Method

plot_difference( pairs; t,x, interpolation,compression, surface,velocity,fourier, label )

Same as plot_difference! but generates and returns the plot.

WaterWaves1D.plot_solution! — Method

plot_solution!( plt; problems; t,x,interpolation,compression,surface,velocity,fourier,label )

Plots in plt the solution of initial-value problems at a given time.

Argument

problems is either an element or a collection (vector, list, etc.) of elements of type Problem.

Keyword arguments (all optional)

t is the time. If not provided, then the last computed time is plotted.

if a vector x is provided and if possible, the solution is interpolated to the collocation points x.

if interpolation is provided as an integer, the solution is interpolated on as many collocation points (if true, then the value 2^3 is chosen, default is false).

if compression is provided as an integer m, only one in m points are plotted (if true, then the value 2^3 is chosen, default is false).

surface, velocity and fourier (booleans) determine respectively whether surface deformation, η, tangential velocity, v, and the Fourier coefficients of η (in log-scale) are plotted.

label defines the label(s) if provided (otherwise labels are inferred from the problems)

WaterWaves1D.plot_solution — Method

plot_solution( problems; t,x, interpolation,compression, surface,velocity,fourier, label )

Same as plot_solution! but generates and returns the plot.

WaterWaves1D.solution — Method

solution(p::Problem;t,x,interpolation)

Gives the solution of a solved initial-value at a given time t.

Arguments

Argument p is of type Problem.

Keyword argument t is optional, the last computed time is returned by default.

Keyword argument x is optional, if provided the solution is interpolated to the collocation vector x.

Keyword argument interpolation is optional, if an integer is provided the solution is interpolated on as many collocation points (if true, then the default value 2^3 is chosen).

Return values

Provides (η,v,x,t) where

η is the surface deformation at collocation points;

v is the tangential velocity (derivative of the trace of the velocity potential) at collocation points;

x is the vector of collocation points;

t the time (first computed time greater or equal to provided t).

WaterWaves1D.solve! — Method

solve!( problems; verbose=true )

Solve (i.e. integrate in time) a collection of initial-value problems.

The argument problems should be a collection (list, array...) of elements of type Problem.

Information are not printed if keyword argument verbose = false (default is true).