Code basics

Abstract types

TimeSolver (RK4, Euler, etc),

AbstractModel (Cheng, Matsuno, etc),

InitialData (Bump, SolitaryWave, etc)

Instances are created from Parameters type.

Missing docstring.

Missing docstring for Parameters. Check Documenter's build log for details.

Une structure Problem représente un problème donné que l'on va résoudre. Les données seront stockées dans data, qui est vide initialement.

ShallowWaterModels.Problem — Type

Problem( model, initial, param; solver)

Builds an initial-value problem which can then be solved (integrated in time) through solve!( problem )

Arguments

model :: AbstractModel, the system of equation solved. May be built, e.g., by WhithamGreenNaghdi(param);

initial :: InitialData, the initial data. May be buit, e.g., by Init(η,v) where η is the surface deformation and v the derivative of the trace of the velocity potential at the surface

param :: NamedTuple, must contain values for

N, the number of collocation points of the spatial grid

L, the half-length of the spatial grid

T, the final time of integration

dt, the timestep

nr (optional, default = T/dt) the number of stored data

solver :: TimeSolver (optional, default = explicit Runge-Kutta fourth order solver), the solver for time integration. May be built, e.g., by RK4(param) or RK4_naive()

Initial data

Missing docstring.

Missing docstring for Bump. Check Documenter's build log for details.

Shallow water models

Missing docstring.

Missing docstring for Cheng. Check Documenter's build log for details.

ShallowWaterModels.Matsuno — Type

Matsuno(params)

ShallowWaterModels.solve! — Method

solve!( problem )

Solves (i.e. integrates in time) an initial-value problem

The argument problem should be of type :: Problem. It may be buit, e.g., by Problem(model, initial, param)

Main program

epsilon = 1/2
N       = 2^12
L       = 10
T       = 5
dt      = 0.001

param = Parameters(epsilon,N,L,T,dt)

problems = [Problem(Cheng,Bump,param,RK4),Problem(Matsuno,Bump,param,RK4)]

Cheng est le modèle utilisé. Il prend en valeur Bump, param et définit 3 fonction:

init qui construit une donnée initiale à partir de Bump : Uinit=init(Bump,param)

Fwave utilisée pour résoudre dt U= Fwave(U) (avec donnée initiale Uinit)

build qui reconstruit la donnée finale (c'est l'application inverse de init). Ufin=final(U,param)

Missing docstring.

Missing docstring for init(::Cheng). Check Documenter's build log for details.

Missing docstring.

Missing docstring for init(::Matsuno). Check Documenter's build log for details.

Missing docstring.

Missing docstring for build(::Cheng). Check Documenter's build log for details.

Missing docstring.

Missing docstring for build(::Matsuno). Check Documenter's build log for details.

simuls = [(Cheng,Bump,param1,RK4),(Cheng,Bump,param2,RK4)]

RK4 solver

ShallowWaterModels.RK4 — Type

`RK4(param,model;k)`

Constructs an object of type `TimeSolver` to be used in `Problem(model, initial, param; solver::TimeSolver)`

param ::NamedTuple should contain a value N (number of collocation points)

model ::AbstractModel is optional and determines the number of equations solved

k=2 ::Int is optional (default = 2) and determines the number of equations solved

for p in problems
    solve!(p)
end