Code basics

Abstract types

TimeSolver (RK4, Euler, etc),

AbstractModel (Cheng, Matsuno, etc),

InitialData (Bump, SolitaryWave, etc)

Instances are created from Parameters type.

Missing docstring.

Missing docstring for Parameters. Check Documenter's build log for details.

Une structure Problem représente un problème donné que l'on va résoudre. Les données seront stockées dans data, qui est vide initialement.

ShallowWaterModels.Problem — Type

Problem( model, initial, param ; verbose=true)

or Problem( model, initial, param, solver ; verbose=true)

Builds an initial-value problem which can then be solved (integrated in time) through solve!( problem )

Arguments

model :: AbstractModel, the system of equation solved.

May be built, e.g., by WaterWaves(param);

initial :: InitialData, the initial data.

May be buit, e.g., by Init(η,v) where η is the surface deformation and v the derivative of the trace of the velocity potential at the surface;

param :: NamedTuple, must contain values for

N, the number of collocation points of the spatial grid

L, the half-length of the spatial grid

T, the final time of integration

dt, the timestep

additionally, it may contain Ns the number of computed data or ns for storing data every ns computation steps (by default, every computed data is stored).

solver :: TimeSolver, the solver for time integration (optional, default is explicit Runge-Kutta fourth order solver).

May be built, e.g., by RK4(model) or RK4_naive().

Information are not printed if keyword verbose = false (default is true).

Initial data

Missing docstring.

Missing docstring for Bump. Check Documenter's build log for details.

Shallow water models

Missing docstring.

Missing docstring for Cheng. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Matsuno. Check Documenter's build log for details.

ShallowWaterModels.solve! — Method

solve!( problem; verbose=true )

Solves (i.e. integrates in time) an initial-value problem

The argument problem should be of type Problem. It may be buit, e.g., by Problem(model, initial, param)

Information are not printed if keyword verbose = false (default is true).

Main program

epsilon = 1/2
N       = 2^12
L       = 10
T       = 5
dt      = 0.001

param = Parameters(epsilon,N,L,T,dt)

problems = [Problem(Cheng,Bump,param,RK4),Problem(Matsuno,Bump,param,RK4)]

Cheng est le modèle utilisé. Il prend en valeur Bump, param et définit 3 fonction:

init qui construit une donnée initiale à partir de Bump : Uinit=init(Bump,param)

Fwave utilisée pour résoudre dt U= Fwave(U) (avec donnée initiale Uinit)

build qui reconstruit la donnée finale (c'est l'application inverse de init). Ufin=final(U,param)

Missing docstring.

Missing docstring for init(::Cheng). Check Documenter's build log for details.

Missing docstring.

Missing docstring for init(::Matsuno). Check Documenter's build log for details.

Missing docstring.

Missing docstring for build(::Cheng). Check Documenter's build log for details.

Missing docstring.

Missing docstring for build(::Matsuno). Check Documenter's build log for details.

simuls = [(Cheng,Bump,param1,RK4),(Cheng,Bump,param2,RK4)]

RK4 solver

ShallowWaterModels.RK4 — Type

RK4(arguments;realdata)

Runge-Kutta fourth order solver.

Constructs an object of type TimeSolver to be used in Problem(model, initial, param; solver::TimeSolver)

Arguments can be either

an object of type AbstractModel;

an Array of size (N,m) where N is the number of collocation points and m the number of data (equations solved);

a Tuple (N,m) as above;

an integer N and an integer m as above (the latter is optional, by default m=2).

a NamedTuple containing a key N and an integer m (the latter is optional, by default m=2).

The keyword argument realdata is optional, and determines whether pre-allocated vectors are real- or complex-valued. By default, they are either determined by the model in case 1., complex-valued otherwise.

for p in problems
    solve!(p)
end