Graphics

WaterWaves1D.create_animation — Method

create_animation( problems; name::String, kwargs... )

Create an animation showing the evolution of initial-value problems.

Argument problems is either an element or a collection (vector, list, etc.) of elements of type Problem.

The animation is saved as name.gif if name is provided.

Other keyword arguments are as follows

xlims allows to specifies the x axis limits for the surface deformation. If nothing is provided (default), then the full numerical basin is represented.

ylims allows to specifies the y axis limits for the surface deformation. If nothing is provided (default), then the limits are determined from the initial data. If anything but a Tuple is provided, the axis limits evolve with the solution.

vlims and flims are as above, but for the velocity and Fourier coefficients plots.

Nframes gives the (maximal) number of frames in the animation.

other arguments of plot_solution!

Return anim, an animation, which can then generate (for instance) a gif through gif(anim, "my_name.gif", fps=15).

WaterWaves1D.plot_difference! — Method

plot_difference!( plt, problems ; t,x,interpolation,compression,fast,surface,velocity,fourier,label )

Plots in plt the difference between solutions of initial-value problems at a given time.

Argument

pairs is a either a collection (vector, list, etc.) of elements of type Problem, or of pairs of such elements. If pairs are provided, then solutions of such problems are compared. If a collection of problems are provided, then all possible pairs are compared.

Keyword arguments (all optional)

t is the time. If not provided, then the last computed time is plotted.

if a vector x is provided and if possible, the solution is interpolated to the collocation points x.

if interpolation is provided as an integer, the solution is interpolated on as many collocation points (if true, then the value 2^3 is chosen, default is false).

if fast = false (default is false), then the second problem is interpolated to the collocation points of the first problem.

if compression is provided as an integer m, only one in m points are plotted (if true, then the value 2^3 is chosen, default is false).

surface, velocity and fourier (booleans) determine respectively whether surface deformation, η, tangential velocity, v, and the Fourier coefficients of η (in log-scale) are plotted.

label defines the label(s) if provided (otherwise labels are inferred from the problems).

WaterWaves1D.plot_difference — Method

plot_difference( pairs; t,x, interpolation,compression, surface,velocity,fourier, label )

Same as plot_difference! but generates and returns the plot.

WaterWaves1D.plot_solution! — Method

plot_solution!( plt; problems; t,x,interpolation,compression,surface,velocity,fourier,label )

Plots in plt the solution of initial-value problems at a given time.

Argument

problems is either an element or a collection (vector, list, etc.) of elements of type Problem.

Keyword arguments (all optional)

t is the time. If not provided, then the last computed time is plotted.

if a vector x is provided and if possible, the solution is interpolated to the collocation points x.

if interpolation is provided as an integer, the solution is interpolated on as many collocation points (if true, then the value 2^3 is chosen, default is false).

if compression is provided as an integer m, only one in m points are plotted (if true, then the value 2^3 is chosen, default is false).

surface, velocity and fourier (booleans) determine respectively whether surface deformation, η, tangential velocity, v, and the Fourier coefficients of η (in log-scale) are plotted.

label defines the label(s) if provided (otherwise labels are inferred from the problems)

WaterWaves1D.plot_solution — Method

plot_solution( problems; t,x, interpolation,compression, surface,velocity,fourier, label )

Same as plot_solution! but generates and returns the plot.