Models

WaterWaves1D.Boussinesq — Type

Boussinesq(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for abcd-Boussinesq models (with b=d and c=0). See Bona, Chen, and Saut

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

two parameters a (default is -1/3) and b (default is +1/3) which determine the model solved. You need a+2*b=1/3 for validity as a long wave model (without surface tension).

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Boussinesq");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function Boussinesq.f! to be called in explicit time-integration solvers;

a function Boussinesq.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function Boussinesq.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.SolitaryWB — Type

SolitaryWB(param; kwargs)

Build the initial data associated with SolitaryWaveWhithamBoussinesq(param; kwargs), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

SolitaryWB(c; ϵ=1,μ=1,N=2^12,kwargs)

Build the initial data with velocity c, dimensionless parameters ϵ and μ, and number of collocation points N, and kwargs the other (optional) keyword arguments as above.

WaterWaves1D.WhithamBoussinesq — Type

WhithamBoussinesq(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for a Boussinesq-type model with full-dispersion property.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

Boussinesq: if true (default is false), compute the standard Boussinesq system instead (see Boussinesq(param;kwargs));

a parameter α which determines the model solved:

If α = 1 (default), then the model has been introduced in Dinvay, Dutykh and Kalisch;

If α = 1/2, then the model is a quasilinear version;

If α < 1/2, then expect instabilities stemming from ill-posedness of the model.

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Whitham-Boussinesq");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WhithamBoussinesq.f! to be called in explicit time-integration solvers;

a function WhithamBoussinesq.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed.

a function WhithamBoussinesq.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.DeepQuadratic — Type

DeepQuadratic(param;dealias,label)

Define an object of type AbstractModel in view of solving the initial-value problem for the quadratic deep-water model proposed by Akers and Milewski and Cheng, Granero-Belinchón, Shkoller and Milewski

Arguments

param is of type NamedTuple and must contain

the dimensionless parameters ϵ (nonlinearity);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

dealias: dealiasing with 1/3 Orlicz rule if true or no dealiasing if false (by default);

label: a label for future references (default is "deep quadratic");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function DeepQuadratic.f! to be called in explicit time-integration solvers;

a function DeepQuadratic.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function DeepQuadratic.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is given by ∂t η = - ∂x v.

WaterWaves1D.DeepQuadratic_fast — Type

DeepQuadratic_fast(param;dealias,label)

Same as DeepQuadratic, but faster.

WaterWaves1D.IsobeKakinuma — Type

IsobeKakinuma(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the Isobe-Kakinuma model.

Argument

param is of type NamedTuple (or a collection NamedTuples) of and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Isobe-Kakinuma");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function IsobeKakinuma.f! to be called in explicit time-integration solvers;

a function IsobeKakinuma.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function IsobeKakinuma.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function IsobeKakinuma.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,Φ), where

Φ is the Vector of the basis functions ϕi (i∈{0,...,N}).

WaterWaves1D.Matsuno — Type

Matsuno(param;dealias,label)

Define an object of type AbstractModel in view of solving the initial-value problem for the quadratic deep-water model proposed by Matsuno.

Arguments

param is of type NamedTuple and must contain

the dimensionless parameters ϵ (nonlinearity);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

dealias: dealiasing with 1/3 Orlicz rule if true or no dealiasing if false (by default);

label: a label for future references (default is "Matsuno");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function DeepQuadratic.f! to be called in explicit time-integration solvers;

a function DeepQuadratic.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function DeepQuadratic.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is a velocity variable which is not the derivative of the trace of the velocity potential (if not null).

WaterWaves1D.Matsuno_fast — Type

Matsuno_fast(param;dealias,label)

Same as Matsuno, but faster.

WaterWaves1D.modifiedMatsuno — Type

modifiedMatsuno(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the modified Matsuno model

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

ν: shallow/deep water multiplication factor. By default, ν=1 if μ≦1 and ν=1/√μ otherwise. Set the infinite-layer case if ν=0 (or μ=Inf).

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "modified Matsuno");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function modifiedMatsuno.f! to be called in explicit time-integration solvers;

a function modifiedMatsuno.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function modifiedMatsuno.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.NonHydrostatic — Type

NonHydrostatic(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the "Non-hydrostatic" model proposed by Bristeau, Mangeney, Sainte-Marie and Seguin

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "non-hydrostatic");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function NonHydrostatic.f! to be called in explicit time-integration solvers;

a function NonHydrostatic.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function NonHydrostatic.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function NonHydrostatic.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.CnoidalSGN — Type

CnoidalSGN(param; P=1)

Build the initial data associated with CnoidalWaveSerreGreenNaghdi(param; P=1), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

WaterWaves1D.SerreGreenNaghdi — Type

SerreGreenNaghdi(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the Serre-Green-Naghdi model

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Green-Naghdi");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function SerreGreenNaghdi.f! to be called in explicit time-integration solvers;

a function SerreGreenNaghdi.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function SerreGreenNaghdi.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function SerreGreenNaghdi.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.SolitarySGN — Type

SolitarySGN(param; x₀=0)

Build the initial data associated with SolitaryWaveSerreGreenNaghdi(param; x₀=0), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

SolitarySGN(c; ϵ=1,μ=1,x₀=0,N=2^12)

Build the initial data with velocity c, center x₀, dimensionless parameters ϵ and μ, and number of collocation points N.

WaterWaves1D.SquareRootDepth — Type

SquareRootDepth(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the "√D" model proposed by Cotter, Holm and Percival

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "square-root depth");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function SquareRootDepth.f! to be called in explicit time-integration solvers;

a function SquareRootDepth.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function SquareRootDepth.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function SquareRootDepth.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.

WaterWaves1D.WWn — Type

WWn(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the water waves expansion proposed by Dommermuth and Yue (doi:10.1017/s002211208700288x), West et al. (doi:10.1029/jc092ic11p11803), Craig and Sulem (doi:10.1006/jcph.1993.1164) (see also the account by Choi : Fifth-order nonlinear spectral model for surface gravity waves: From pseudo-spectral to spectral formulations, RIMS Kokyuroku, 2019) with the "rectification" method proposed by Duchêne and Melinand.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param)

Optional keyword arguments

ν: shallow/deep water multiplication factor. By default, ν=1 if μ≦1 and ν=1/√μ otherwise. Set the infinite-layer case if ν=0 (or μ=Inf).

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

n :: Int: the order of the expansion; linear system if 1, quadratic if 2, cubic if 3, quartic if 4 (default and other values yield 2);

δ and m: parameters of the rectifier operator, set as k->min(1,|δ*k|^m) or k->min(1,|δ*k|^m[1]*exp(1-|δ*k|^m[2])) if m is a couple

(by default is δ=0, i.e. no regularization and m=-1. Notice m=-Inf and δ>0 yields a cut-off filter);

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "WWn" with n the order of the expansion);

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WWn.f! to be called in explicit time-integration solvers (also WWn.f1! and WWn.f2! for the symplectic Euler solver);

a function WWn.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WWn.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential.

WaterWaves1D.WaterWaves — Type

WaterWaves(param; kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the water waves system (via conformal mapping, see Zakharov, Dyachenko and Vasilyev).

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

ν: shallow/deep water multiplication factor (see Lannes, The water waves problem). By default, ν=1 if μ≦1 and ν=1/√μ otherwise.

IL: Set the infinite-layer case if IL=true (or μ=Inf, or ν=0), in which case ϵ is the steepness parameter. Default is false.

method ∈ {1,2,3}: method used to initialize the conformal mapping, as a fix-point problem F(u)=u

if method == 1, use standard contraction fix-point iteration;

if method == 2, use Newton algorithm with GMRES iterative solver to invert the Jacobian;

if method == 3, use Newton algorithm with direct solver to invert the Jacobian;

tol: (relative) tolerance of the fix-point algorithm (default is 1e-16);

maxiter: the maximal number of iteration in the fix-point algorithm (default is 100);

ktol: tolerance of the low-pass Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "water waves");

verbose: prints information if true (default is true).

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WaterWaves.f! to be called in the explicit time-integration solver (also WaterWaves.f1! and WaterWaves.f2! for the symplectic Euler solver);

a function WaterWaves.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WaterWaves.mapfro which from such data matrix returns the Tuple of real vectors (x,η,v), where

x is a vector of collocation points (non-regularly spaced);

η is the surface deformation at points x;

v is the derivative of the trace of the velocity potential at points x.

WaterWaves1D.SolitaryWGN — Type

SolitaryWGN(param; kwargs)

Build the initial data associated with SolitaryWaveWhithamGreenNaghdi(param; kwargs), of type InitialData, to be used in initial-value problems Problem(model, initial::InitialData, param).

SolitaryWGN(c; ϵ=1,μ=1,N=2^12,kwargs)

Build the initial data with velocity c, dimensionless parameters ϵ and μ, and number of collocation points N, and kwargs the other (optional) keyword arguments as above.

WaterWaves1D.WhithamGreenNaghdi — Type

WhithamGreenNaghdi(param;kwargs)

Define an object of type AbstractModel in view of solving the initial-value problem for the fully dispersive Green-Naghdi model proposed by Duchêne, Israwi and Talhouk.

Argument

param is of type NamedTuple and must contain

dimensionless parameters ϵ (nonlinearity) and μ (dispersion);

numerical parameters to construct the mesh of collocation points as mesh = Mesh(param).

Optional keyword arguments

SGN: if true (default is false), compute the Serre-Green-Naghdi (SGN) instead of Whitham-Green-Naghdi (WGN) system (see SerreGreenNaghdi(param;kwargs));

iterative: solve the elliptic problem through GMRES if true, LU decomposition if false (default is true);

precond: use a (left) preconditioner for GMRES if true (default), choose precond as the preconditioner if provided;

gtol: relative tolerance of the GMRES algorithm (default is 1e-14);

restart: the corresponding option of the GMRES algorithm (default is 100);

maxiter: the corresponding option of GMRES (default is nothing);

ktol: tolerance of the Krasny filter (default is 0, i.e. no filtering);

dealias: dealiasing with Orlicz rule 1-dealias/(dealias+2) (default is 0, i.e. no dealiasing);

label: a label for future references (default is "Whitham-Green-Naghdi");

Return values

Generate necessary ingredients for solving an initial-value problem via solve!:

a function WhithamGreenNaghdi.f! to be called in explicit time-integration solvers;

a function WhithamGreenNaghdi.mapto which from (η,v) of type InitialData provides the raw data matrix on which computations are to be executed;

a function WhithamGreenNaghdi.mapfro which from such data matrix returns the Tuple of real vectors (η,v), where

η is the surface deformation;

v is the derivative of the trace of the velocity potential;

additionally, a handy function WhithamGreenNaghdi.mapfrofull which from data matrix returns the Tuple of real vectors (η,v,u), where

u corresponds to the layer-averaged velocity.